requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
alembic=1.16.4=pypi_0
amqp=5.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
bcrypt=4.3.0=pypi_0
billiard=4.2.1=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
celery=5.5.3=pypi_0
certifi=2025.7.14=pypi_0
charset-normalizer=3.4.2=pypi_0
click=8.2.1=pypi_0
click-didyoumean=0.3.1=pypi_0
click-plugins=1.1.1.2=pypi_0
click-repl=0.3.0=pypi_0
colorama=0.4.6=pypi_0
dnspython=2.7.0=pypi_0
docopt=0.6.2=pypi_0
email-validator=2.2.0=pypi_0
eventlet=0.40.2=pypi_0
expat=2.7.1=h8ddb27b_0
fastapi=0.116.1=pypi_0
fastapi-cli=0.0.8=pypi_0
fastapi-cloud-cli=0.1.4=pypi_0
flower=2.0.1=pypi_0
greenlet=3.2.3=pypi_0
h11=0.16.0=pypi_0
httpcore=1.0.9=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
humanize=4.12.3=pypi_0
idna=3.10=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.6=pypi_0
kombu=5.5.4=pypi_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
mako=1.3.10=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
mdurl=0.1.2=pypi_0
openssl=3.0.17=h35632f6_0
orjson=3.11.0=pypi_0
packaging=25.0=pypi_0
passlib=1.7.4=pypi_0
pip=25.1=pyhc872135_2
pipreqs=0.4.13=pypi_0
prometheus-client=0.22.1=pypi_0
prompt-toolkit=3.0.51=pypi_0
pydantic=2.11.7=py313haa95532_0
pydantic-core=2.33.2=py313h215eeae_0
pydantic-extra-types=2.10.5=pypi_0
pydantic-settings=2.10.1=pypi_0
pygments=2.19.2=pypi_0
pyjwt=2.10.1=pypi_0
pymysql=1.1.1=pypi_0
pyreadline3=3.5.4=pypi_0
python=3.13.5=h286a616_100_cp313
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.1.1=pypi_0
python-multipart=0.0.20=pypi_0
python_abi=3.13=0_cp313
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
redis=6.2.0=pypi_0
requests=2.32.4=pypi_0
rich=14.0.0=pypi_0
rich-toolkit=0.14.8=pypi_0
rignore=0.6.4=pypi_0
sentry-sdk=2.33.0=pypi_0
setuptools=78.1.1=py313haa95532_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.41=py313h54f65d0_0
sqlite=3.50.2=hda9a48d_1
sqlmodel=0.0.24=py313haa95532_0
starlette=0.47.2=pypi_0
tk=8.6.14=h5e9d12e_1
tornado=6.5.1=pypi_0
typer=0.16.0=pypi_0
typing-extensions=4.14.1=pypi_0
typing-inspection=0.4.1=pypi_0
typing_extensions=4.12.2=py313haa95532_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=haa95532_0
ujson=5.10.0=pypi_0
urllib3=2.5.0=pypi_0
uuid=1.30=pypi_0
uvicorn=0.35.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vine=5.1.0=pypi_0
vs2015_runtime=14.44.35208=ha6b5a95_10
watchfiles=1.1.0=pypi_0
wcwidth=0.2.13=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py313haa95532_0
xz=5.6.4=h4754444_1
yarg=0.1.10=pypi_0
zlib=1.2.13=h8cc25b3_1
